--- src/inputsGenerator/imports/index.ts
@@ -1,4 +1,0 @@
-import { ConfigInternal } from '../../utils/config';
-
-export const getImports = (config: ConfigInternal) =>
-  [config.inputs.prismaImporter, config.inputs.builderImporter].join('\n');
