+++ src/inputsGenerator/utils/parts.test.ts
@@ -1,13 +1,11 @@
 import { getSampleDMMF } from '../../tests/getPrismaSchema';
-import { getDefaultConfig } from '../../utils/config';
-import { getInputs } from './index';
+import { getInputs } from './parts';
 
 describe('getInputs', () => {
   test('should create input', async () => {
     const dmmf = await getSampleDMMF('complex');
     const builtString = `export const UserCreateInput = builder.inputRef<Prisma.UserCreateInput>('UserCreateInput').implement({`;
-    const defaultConfig = getDefaultConfig();
-    const includedInputs = getInputs(defaultConfig, dmmf);
+    const includedInputs = getInputs(dmmf);
     expect(includedInputs.includes(builtString)).toBe(true);
   });
 });
